+++ src/main/java/uk/gov/hmcts/reform/preapi/email/IEmailService.java
@@ -1,7 +1,7 @@
 package uk.gov.hmcts.reform.preapi.email;
 
-import uk.gov.hmcts.reform.preapi.email.govnotify.templates.EditEmailParameters;
 import uk.gov.hmcts.reform.preapi.entities.Case;
+import uk.gov.hmcts.reform.preapi.entities.EditRequest;
 import uk.gov.hmcts.reform.preapi.entities.User;
 
 import java.sql.Timestamp;
@@ -22,5 +22,5 @@ public interface IEmailService {
 
     EmailResponse emailVerification(String email, String firstName, String lastName, String verificationCode);
 
-    Optional<EmailResponse> sendEmailAboutEditingRequest(EditEmailParameters editEmailParameters);
+    Optional<EmailResponse> sendEmailAboutEditingRequest(EditRequest editRequest);
 }
