+++ src/test/java/uk/gov/hmcts/reform/preapi/email/govnotify/GovNotifyTest.java
@@ -1,0 +1,530 @@
+package uk.gov.hmcts.reform.preapi.email.govnotify;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.mockito.ArgumentCaptor;
+import org.mockito.Captor;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.TestPropertySource;
+import org.springframework.test.context.bean.override.mockito.MockitoBean;
+import uk.gov.hmcts.reform.preapi.email.EmailResponse;
+import uk.gov.hmcts.reform.preapi.email.EmailServiceFactory;
+import uk.gov.hmcts.reform.preapi.email.govnotify.templates.CaseClosed;
+import uk.gov.hmcts.reform.preapi.email.govnotify.templates.CaseClosureCancelled;
+import uk.gov.hmcts.reform.preapi.email.govnotify.templates.CasePendingClosure;
+import uk.gov.hmcts.reform.preapi.email.govnotify.templates.PortalInvite;
+import uk.gov.hmcts.reform.preapi.email.govnotify.templates.RecordingEdited;
+import uk.gov.hmcts.reform.preapi.email.govnotify.templates.RecordingReady;
+import uk.gov.hmcts.reform.preapi.entities.Case;
+import uk.gov.hmcts.reform.preapi.entities.Court;
+import uk.gov.hmcts.reform.preapi.entities.User;
+import uk.gov.hmcts.reform.preapi.exception.EmailFailedToSendException;
+import uk.gov.service.notify.NotificationClient;
+import uk.gov.service.notify.NotificationClientException;
+import uk.gov.service.notify.SendEmailResponse;
+
+import java.sql.Timestamp;
+import java.util.List;
+import java.util.Map;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
+
+@SpringBootTest(classes = GovNotify.class)
+@TestPropertySource(properties = {
+    "email.govNotify.key=test",
+    "portal.url=http://localhost:8080"
+})
+public class GovNotifyTest {
+
+    @MockitoBean
+    NotificationClient mockGovNotifyClient;
+
+    private static final String govNotifyEmailResponse = """
+        {
+          "id": "740e5834-3a29-46b4-9a6f-16142fde533a",
+          "reference": "STRING",
+          "content": {
+            "subject": "SUBJECT TEXT",
+            "body": "MESSAGE TEXT",
+            "from_email": "SENDER EMAIL"
+          },
+          "uri": "https://api.notifications.service.gov.uk/v2/notifications/740e5834-3a29-46b4-9a6f-16142fde533a",
+          "template": {
+            "id": "f33517ff-2a88-4f6e-b855-c550268ce08a",
+            "version": 1,
+            "uri": "https://api.notifications.service.gov.uk/v2/template/f33517ff-2a88-4f6e-b855-c550268ce08a"
+          }
+        }""";
+
+
+    private GovNotify underTest;
+
+    @BeforeEach
+    void setUp() throws NotificationClientException {
+        when(mockGovNotifyClient.sendEmail(any(), any(), any(), any()))
+            .thenReturn(new SendEmailResponse(govNotifyEmailResponse));
+
+        underTest = new GovNotify("http://localhost:8080", mockGovNotifyClient);
+    }
+
+    @DisplayName("Should create CaseClosed template")
+    @Test
+    void shouldCreateCaseClosedTemplate() {
+        var template = new CaseClosed("to", "firstName", "lastName", "caseRef");
+        assertThat(template.getTemplateId()).isEqualTo("ee5c6d3f-e934-4053-9f48-0ba082b8caf4");
+    }
+
+    @DisplayName("Should create RecordingEdited template")
+    @Test
+    void shouldCreateRecordingEditedTemplate() {
+        var template = new RecordingEdited("to", "firstName", "lastName", "caseRef", "courtName", "portalLink");
+        assertThat(template.getTemplateId()).isEqualTo("1da03824-84e8-425d-b913-c2bac661e64a");
+    }
+
+    @DisplayName("Should create CaseClosureCancelled template")
+    @Test
+    void shouldCreateCaseClosureCancelledTemplate() {
+        var template = new CaseClosureCancelled("to", "firstName", "lastName", "caseRef");
+        assertThat(template.getTemplateId()).isEqualTo("5fba3021-a835-4f98-a575-83cc5fcb83a4");
+    }
+
+    @DisplayName("Should create RecordingReady template")
+    @Test
+    void shouldCreateRecordingReadyTemplate() {
+        var template = new RecordingReady("to", "firstName", "lastName", "caseRef", "courtName", "portalLink");
+        assertThat(template.getTemplateId()).isEqualTo("6ad8d468-4a18-4180-9c08-c6fae055a385");
+    }
+
+    @DisplayName("Should create CasePendingClosure template")
+    @Test
+    void shouldCreateCasePendingClosureTemplate() {
+        var template = new CasePendingClosure(
+            "to",
+            "firstName",
+            "lastName",
+            "caseRef",
+            Timestamp.valueOf("2025-01-01 00:00:00.0")
+        );
+        assertThat(template.getTemplateId()).isEqualTo("5322ba5c-f4c4-4d1b-807c-16f56f0d8d0c");
+    }
+
+    @DisplayName("Should create PortalInvite template")
+    @Test
+    void shouldCreatePortalInviteTemplate() {
+        var template = new PortalInvite(
+            "to",
+            "firstName",
+            "lastName",
+            "portalUrl",
+            "guideLink",
+            "processGuideLink",
+            "faqsLink",
+            "editingRequestForm"
+        );
+        assertThat(template.getTemplateId()).isEqualTo("e04adfb8-58e0-44be-ab42-bd6d896ccfb7");
+    }
+
+    @DisplayName("Should create EmailResponse from GovNotify response")
+    @Test
+    void shouldCreateEmailResponseFromGovNotifyResponse() {
+        var response = new SendEmailResponse(govNotifyEmailResponse);
+        var emailResponse = EmailResponse.fromGovNotifyResponse(response);
+        assertThat(emailResponse.getFromEmail()).isEqualTo("SENDER EMAIL");
+        assertThat(emailResponse.getSubject()).isEqualTo("SUBJECT TEXT");
+        assertThat(emailResponse.getBody()).isEqualTo("MESSAGE TEXT");
+    }
+
+    @DisplayName(("Email service factory"))
+    @Test
+    void shouldCreateEmailServiceFactory() {
+        var govNotify = new GovNotify("GovNotify", mockGovNotifyClient);
+        var emailServiceFactory = new EmailServiceFactory("GovNotify", true, List.of(govNotify));
+        assertThat(emailServiceFactory).isNotNull();
+        assertThat(emailServiceFactory.getEnabledEmailService()).isEqualTo(govNotify);
+        assertThat(emailServiceFactory.isEnabled()).isTrue();
+
+        assertThat(emailServiceFactory.getEnabledEmailService("GovNotify")).isEqualTo(govNotify);
+        assertThrows(IllegalArgumentException.class, () -> emailServiceFactory.getEnabledEmailService("nonexistent"));
+
+        assertThrows(
+            IllegalArgumentException.class, () ->
+                new EmailServiceFactory("nonexistent", true, List.of(govNotify))
+        );
+    }
+
+
+    @DisplayName(("Should send recording ready email"))
+    @Test
+    void shouldSendRecordingReadyEmail() throws NotificationClientException {
+        var response = underTest.recordingReady(getSampleUser(), getSampleCase());
+
+        assertThat(response.getFromEmail()).isEqualTo("SENDER EMAIL");
+        assertThat(response.getSubject()).isEqualTo("SUBJECT TEXT");
+        assertThat(response.getBody()).isEqualTo("MESSAGE TEXT");
+    }
+
+    @DisplayName(("Should send recording edited email"))
+    @Test
+    void shouldSendRecordingEditedEmail() {
+        var response = underTest.recordingEdited(getSampleUser(), getSampleCase());
+
+        assertThat(response.getFromEmail()).isEqualTo("SENDER EMAIL");
+        assertThat(response.getSubject()).isEqualTo("SUBJECT TEXT");
+        assertThat(response.getBody()).isEqualTo("MESSAGE TEXT");
+    }
+
+    @DisplayName(("Should send portal invite email"))
+    @Test
+    void shouldSendPortalInviteEmail() throws NotificationClientException {
+        var response = underTest.portalInvite(getSampleUser());
+
+        assertThat(response.getFromEmail()).isEqualTo("SENDER EMAIL");
+        assertThat(response.getSubject()).isEqualTo("SUBJECT TEXT");
+        assertThat(response.getBody()).isEqualTo("MESSAGE TEXT");
+    }
+
+    @DisplayName(("Should send case pending closure email"))
+    @Test
+    void shouldSendCasePendingClosureEmail() throws NotificationClientException {
+        var response = underTest.casePendingClosure(getSampleUser(), getSampleCase(), Timestamp.valueOf("2025-01-01 00:00:00.0"));
+
+        assertThat(response.getFromEmail()).isEqualTo("SENDER EMAIL");
+        assertThat(response.getSubject()).isEqualTo("SUBJECT TEXT");
+        assertThat(response.getBody()).isEqualTo("MESSAGE TEXT");
+    }
+
+    @DisplayName(("Should send case closed email"))
+    @Test
+    void shouldSendCaseClosedEmail() throws NotificationClientException {
+        var response = underTest.caseClosed(getSampleUser(), getSampleCase());
+
+        assertThat(response.getFromEmail()).isEqualTo("SENDER EMAIL");
+        assertThat(response.getSubject()).isEqualTo("SUBJECT TEXT");
+        assertThat(response.getBody()).isEqualTo("MESSAGE TEXT");
+    }
+
+    @DisplayName(("Should send case closure cancelled email"))
+    @Test
+    void shouldSendCaseClosureCancelledEmail() throws NotificationClientException {
+        var response = underTest.caseClosureCancelled(getSampleUser(), getSampleCase());
+
+        assertThat(response.getFromEmail()).isEqualTo("SENDER EMAIL");
+        assertThat(response.getSubject()).isEqualTo("SUBJECT TEXT");
+        assertThat(response.getBody()).isEqualTo("MESSAGE TEXT");
+    }
+
+    @Captor
+    private ArgumentCaptor<Map<String, Object>> variablesCaptor;
+
+//    @Test
+//    @DisplayName("Should send editing rejection email")
+//    void sendEditRejectionEmail() throws NotificationClientException {
+//        EditRequest mockEditRequest = mock(EditRequest.class);
+//        when()
+//
+//
+//        EmailResponse response = underTest.sendEmailAboutEditingRequest(params)
+//            .orElseThrow(() -> new EmailFailedToSendException("Something went wrong"));
+//
+//        assertThat(response.getFromEmail()).isEqualTo("SENDER EMAIL");
+//        assertThat(response.getSubject()).isEqualTo("SUBJECT TEXT");
+//        assertThat(response.getBody()).isEqualTo("MESSAGE TEXT");
+//
+//        ArgumentCaptor<String> emailAddressCaptor = ArgumentCaptor.forClass(String.class);
+//        ArgumentCaptor<String> templateCaptor = ArgumentCaptor.forClass(String.class);
+//        ArgumentCaptor<String> referenceCaptor = ArgumentCaptor.forClass(String.class);
+//
+//        verify(mockGovNotifyClient, times(1)).sendEmail(
+//            templateCaptor.capture(),
+//            emailAddressCaptor.capture(),
+//            variablesCaptor.capture(),
+//            referenceCaptor.capture()
+//        );
+//
+//        // Gov Notify template ID
+//        assertThat(templateCaptor.getValue()).isEqualTo("aa2a836f-b6f0-46dc-91e0-1698822c5137");
+//        assertThat(emailAddressCaptor.getValue()).isEqualTo(params.getToEmailAddress());
+//        assertThat(variablesCaptor.getValue().get("case_reference")).isEqualTo(params.getCaseReference());
+//        assertThat(variablesCaptor.getValue().get("defendant_names")).isEqualTo(params.getDefendantName());
+//        assertThat(variablesCaptor.getValue().get("court_name")).isEqualTo(params.getCourtName());
+//        assertThat(variablesCaptor.getValue().get("edit_summary")).isEqualTo(params.getEditSummary());
+//        assertThat(variablesCaptor.getValue().get("edit_count"))
+//            .isEqualTo(params.getNumberOfRequestedEditInstructions());
+//        assertThat(variablesCaptor.getValue().get("jointly_agreed")).isEqualTo("Yes");
+//        assertThat(variablesCaptor.getValue().get("rejection_reason")).isEqualTo(params.getRejectionReason());
+//        assertThat(variablesCaptor.getValue().get("portal_link")).isEqualTo("http://localhost:8080");
+//    }
+
+    @DisplayName(("Should fail to send recording ready email"))
+    @Test
+    void shouldFailToSendRecordingReadyEmail() throws NotificationClientException {
+
+        when(mockGovNotifyClient.sendEmail(any(), any(), any(), any()))
+            .thenThrow(mock(NotificationClientException.class));
+
+        var message = assertThrows(
+            EmailFailedToSendException.class,
+            () -> underTest.recordingReady(getSampleUser(), getSampleCase())
+        ).getMessage();
+
+        assertThat(message).isEqualTo("Failed to send email to: " + getSampleUser().getEmail());
+    }
+
+    @DisplayName(("Should fail to send recording edited email"))
+    @Test
+    void shouldFailToSendRecordingEditedEmail() throws NotificationClientException {
+
+        when(mockGovNotifyClient.sendEmail(any(), any(), any(), any()))
+            .thenThrow(mock(NotificationClientException.class));
+
+        var message = assertThrows(
+            EmailFailedToSendException.class,
+            () -> underTest.recordingEdited(getSampleUser(), getSampleCase())
+        ).getMessage();
+
+        assertThat(message).isEqualTo("Failed to send email to: " + getSampleUser().getEmail());
+    }
+
+    @DisplayName(("Should fail to send portal invite email"))
+    @Test
+    void shouldFailToSendPortalInviteEmail() throws NotificationClientException {
+
+        when(mockGovNotifyClient.sendEmail(any(), any(), any(), any()))
+            .thenThrow(mock(NotificationClientException.class));
+
+        var message = assertThrows(
+            EmailFailedToSendException.class,
+            () -> underTest.portalInvite(getSampleUser())
+        ).getMessage();
+
+        assertThat(message).isEqualTo("Failed to send email to: " + getSampleUser().getEmail());
+    }
+
+    @DisplayName(("Should fail to send case pending closure email"))
+    @Test
+    void shouldFailToSendCasePendingClosureEmail() throws NotificationClientException {
+
+        when(mockGovNotifyClient.sendEmail(any(), any(), any(), any()))
+            .thenThrow(mock(NotificationClientException.class));
+
+        var message = assertThrows(
+            EmailFailedToSendException.class,
+            () -> underTest.casePendingClosure(
+                getSampleUser(),
+                getSampleCase(),
+                Timestamp.valueOf("2025-01-01 00:00:00.0")
+            )
+        )
+            .getMessage();
+
+        assertThat(message).isEqualTo("Failed to send email to: " + getSampleUser().getEmail());
+    }
+
+    @DisplayName(("Should fail to send case closed email"))
+    @Test
+    void shouldFailToSendCaseClosedEmail() throws NotificationClientException {
+
+        when(mockGovNotifyClient.sendEmail(any(), any(), any(), any()))
+            .thenThrow(mock(NotificationClientException.class));
+
+        var message = assertThrows(
+            EmailFailedToSendException.class,
+            () -> underTest.caseClosed(getSampleUser(), getSampleCase())
+        ).getMessage();
+
+        assertThat(message).isEqualTo("Failed to send email to: " + getSampleUser().getEmail());
+    }
+
+    @DisplayName(("Should fail to send case closure cancelled email"))
+    @Test
+    void shouldFailToSendCaseClosureCancelledEmail() throws NotificationClientException {
+
+        when(mockGovNotifyClient.sendEmail(any(), any(), any(), any()))
+            .thenThrow(mock(NotificationClientException.class));
+
+        var message = assertThrows(
+            EmailFailedToSendException.class,
+            () -> underTest.caseClosureCancelled(getSampleUser(), getSampleCase())
+        ).getMessage();
+
+        assertThat(message).isEqualTo("Failed to send email to: " + getSampleUser().getEmail());
+    }
+
+    @DisplayName(("Should send email verification email"))
+    @Test
+    void shouldSendEmailVerificationEmail() {
+        var user = getSampleUser();
+        var response = underTest.emailVerification(user.getEmail(), user.getFirstName(), user.getLastName(), "123456");
+
+        assertThat(response.getFromEmail()).isEqualTo("SENDER EMAIL");
+        assertThat(response.getSubject()).isEqualTo("SUBJECT TEXT");
+        assertThat(response.getBody()).isEqualTo("MESSAGE TEXT");
+    }
+
+    @DisplayName(("Should fail to send email verification email"))
+    @Test
+    void shouldFailToSendEmailVerificationEmail() throws NotificationClientException {
+        when(mockGovNotifyClient.sendEmail(any(), any(), any(), any()))
+            .thenThrow(mock(NotificationClientException.class));
+
+        var user = getSampleUser();
+        var message = assertThrows(
+            EmailFailedToSendException.class,
+            () -> underTest.emailVerification(
+                user.getEmail(),
+                user.getFirstName(),
+                user.getLastName(),
+                "123456"
+            )
+        ).getMessage();
+
+        assertThat(message).isEqualTo("Failed to send email to: " + getSampleUser().getEmail());
+    }
+
+//    @Test
+//    @DisplayName("Should fail to send editing email")
+//    void shouldFailToSendEditingEmail() throws NotificationClientException {
+//        when(mockGovNotifyClient.sendEmail(any(), any(), any(), any()))
+//            .thenThrow(mock(NotificationClientException.class));
+//
+//        EditEmailParameters params = getSampleEditEmailParams(EditRequestStatus.REJECTED, true);
+//
+//        var message = assertThrows(
+//            EmailFailedToSendException.class,
+//            () -> underTest.sendEmailAboutEditingRequest(params)
+//        )
+//            .getMessage();
+//
+//        assertThat(message).isEqualTo("Failed to send email to: " + getSampleCase().getCourt().getGroupEmail());
+//    }
+
+    @DisplayName("Should prefer alternative email when it ends with .cjsm.net")
+    @Test
+    void shouldPreferAlternativeEmailWithCjsmNet() throws Exception {
+        var user = new User();
+        user.setEmail("user@example.com");
+        user.setAlternativeEmail("user@example.com.cjsm.net");
+
+        var method = GovNotify.class.getDeclaredMethod("getUsersPreferredEmail", User.class);
+        method.setAccessible(true);
+
+        var result = (String) method.invoke(underTest, user);
+
+        assertThat(result).isEqualTo("user@example.com.cjsm.net");
+    }
+
+    @DisplayName("Should prefer primary email when it ends with .cjsm.net and alternative doesn't")
+    @Test
+    void shouldPreferPrimaryEmailWithCjsmNet() throws Exception {
+        var user = new User();
+        user.setEmail("user@example.com.cjsm.net");
+        user.setAlternativeEmail("user@example.com");
+
+        var method = GovNotify.class.getDeclaredMethod("getUsersPreferredEmail", User.class);
+        method.setAccessible(true);
+
+        var result = (String) method.invoke(underTest, user);
+
+        assertThat(result).isEqualTo("user@example.com.cjsm.net");
+    }
+
+    @DisplayName("Should prefer alternative email when both end with .cjsm.net")
+    @Test
+    void shouldPreferAlternativeEmailWhenBothHaveCjsmNet() throws Exception {
+        var user = new User();
+        user.setEmail("user@example1.com.cjsm.net");
+        user.setAlternativeEmail("user@example2.com.cjsm.net");
+
+        var method = GovNotify.class.getDeclaredMethod("getUsersPreferredEmail", User.class);
+        method.setAccessible(true);
+
+        var result = (String) method.invoke(underTest, user);
+
+        assertThat(result).isEqualTo("user@example2.com.cjsm.net");
+    }
+
+    @DisplayName("Should fallback to primary email when neither ends with .cjsm.net")
+    @Test
+    void shouldFallbackToPrimaryEmailWhenNoCjsmNet() throws Exception {
+        var user = new User();
+        user.setEmail("user@example.com");
+        user.setAlternativeEmail("user@other.com");
+
+        var method = GovNotify.class.getDeclaredMethod("getUsersPreferredEmail", User.class);
+        method.setAccessible(true);
+
+        var result = (String) method.invoke(underTest, user);
+
+        assertThat(result).isEqualTo("user@example.com");
+    }
+
+    @DisplayName("Should fallback to primary email when alternative email is null")
+    @Test
+    void shouldFallbackToPrimaryEmailWhenAlternativeIsNull() throws Exception {
+        var user = new User();
+        user.setEmail("user@example.com");
+        user.setAlternativeEmail(null);
+
+        var method = GovNotify.class.getDeclaredMethod("getUsersPreferredEmail", User.class);
+        method.setAccessible(true);
+
+        var result = (String) method.invoke(underTest, user);
+
+        assertThat(result).isEqualTo("user@example.com");
+    }
+
+    @DisplayName("Should prefer primary email when it ends with .cjsm.net and alternative is null")
+    @Test
+    void shouldPreferPrimaryEmailWhenAlternativeIsNullAndPrimaryHasCjsmNet() throws Exception {
+        var user = new User();
+        user.setEmail("user@example.com.cjsm.net");
+        user.setAlternativeEmail(null);
+
+        var method = GovNotify.class.getDeclaredMethod("getUsersPreferredEmail", User.class);
+        method.setAccessible(true);
+
+        var result = (String) method.invoke(underTest, user);
+
+        assertThat(result).isEqualTo("user@example.com.cjsm.net");
+    }
+
+    private static User getSampleUser() {
+        var user = new User();
+        user.setFirstName("John");
+        user.setLastName("Doe");
+        user.setEmail("johndoe@example.com");
+        return user;
+    }
+
+    private static Case getSampleCase() {
+        var forCase = new Case();
+        forCase.setReference("123456");
+        var court = new Court();
+        court.setName("Court Name");
+        court.setGroupEmail("group-email@example.com");
+        forCase.setCourt(court);
+        return forCase;
+    }
+
+//    private static EditEmailParameters getSampleEditEmailParams(EditRequestStatus editRequestStatus,
+//                                                                boolean jointlyAgreed) {
+//        Case testCase = getSampleCase();
+//        return EditEmailParameters.builder()
+//            .toEmailAddress(testCase.getCourt().getGroupEmail())
+//            .caseReference(testCase.getReference())
+//            .courtName(testCase.getCourt().getName())
+//            .witnessName(getSampleUser().getFirstName())
+//            .defendantName(getSampleUser().getFullName())
+//            .editSummary("my edits")
+//            .editRequestStatus(editRequestStatus)
+//            .numberOfRequestedEditInstructions(3)
+//            .jointlyAgreed(jointlyAgreed)
+//            .rejectionReason("REJECTION REASON")
+//            .build();
+//    }
+}
